+++ chess/change-log.md
@@ -1,9 +1,12 @@
 ## Version 1.2.0
+  - Use image files for pieces to remove "DejaVu Sans" font dependency.
+
+## Version 1.2.0
   - Show legal move hints.
 
 ## Version 1.1.0
   - Add drag-and-drop for moving pieces.
 
 ## Version 1.0.0
- - basic functions of a chess game
- - 2 player mode only
+  - basic functions of a chess game
+  - 2 player mode only